AG-13242 - Switch example generator to tinypool to improve parallelism.
