+++ package.json
@@ -142,6 +142,7 @@
       "plugins/ag-charts-generate-code-reference-files",
       "plugins/ag-charts-generate-example-files",
       "plugins/ag-charts-task-autogen",
+      "plugins/plugin-utils",
       "libraries/ag-charts-test",
       "tests/angular-package-tests",
       "tests/react-package-tests",
