+++ plugins/ag-charts-generate-chart-thumbnail/package.json
@@ -13,7 +13,6 @@
     "ag-charts-enterprise": "11.0.3-beta.20250112",
     "ag-charts-test": "11.0.3-beta.20250112",
     "@swc/helpers": "0.5.3",
-    "tinypool": "^1.0.2",
     "process": "0.11.10"
   },
   "type": "commonjs",
