+++ plugins/ag-charts-generate-chart-thumbnail/src/executors/generate/batch-instance.ts
@@ -1,6 +1,6 @@
 import type { ExecutorContext } from '@nx/devkit';
+import type { BatchExecutorTaskResult } from 'plugin-utils';
 
-import type { BatchExecutorTaskResult } from '../../executors-utils';
 import { type ExecutorOptions, generateFiles } from './executor';
 
 export type Message = {
