+++ plugins/ag-charts-generate-chart-thumbnail/src/executors/generate/executor.ts
@@ -3,10 +3,10 @@ import type { ExecutorContext } from '@nx/devkit';
 import { Canvas, CanvasRenderingContext2D } from 'canvas';
 import path from 'path';
 import { Path2D, applyPath2DToCanvasRenderingContext } from 'path2d';
+import { consolePrefix, ensureDirectory, readJSONFile } from 'plugin-utils';
 
 import type { AgChartThemeName } from 'ag-charts-community';
 
-import { consolePrefix, ensureDirectory, readJSONFile } from '../../executors-utils';
 import { generateThumbnail } from './generator/thumbnailGenerator';
 
 global.Path2D = Path2D;
