+++ plugins/ag-charts-generate-example-files/package.json
@@ -12,7 +12,6 @@
     "@swc/helpers": "0.5.3",
     "jscodeshift": "^0.15.0",
     "prettier": "^3.2.5",
-    "tinypool": "^0.8.1",
     "ag-charts-community": "11.0.3-beta.20250112",
     "ag-charts-test": "11.0.3-beta.20250112"
   },
