+++ plugins/ag-charts-generate-example-files/project.json
@@ -6,7 +6,7 @@
   "targets": {
     "build": {
       "executor": "@nx/js:swc",
-      "dependsOn": ["ag-charts-community:build:types", "ag-charts-community:build:package"],
+      "dependsOn": ["ag-charts-community:build:types", "ag-charts-community:build:package", "plugin-utils:build"],
       "inputs": ["{projectRoot}/src/**"],
       "outputs": ["{options.outputPath}"],
       "cache": true,
