+++ plugins/ag-charts-generate-example-files/src/executors/generate/generator/examplesGenerator.ts
@@ -1,7 +1,7 @@
 import fs from 'fs/promises';
 import path from 'path';
+import { readFile } from 'plugin-utils';
 
-import { readFile } from '../../../executors-utils';
 import { ANGULAR_GENERATED_MAIN_FILE_NAME, SOURCE_ENTRY_FILE_NAME } from './constants';
 import { transformPlainEntryFile } from './transformPlainEntryFile';
 import chartVanillaSrcParser from './transformation-scripts/chart-vanilla-src-parser';
