+++ plugins/plugin-utils/README.md
@@ -1,0 +1,7 @@
+# plugin-utils
+
+This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build plugin-utils` to build the library.
