+++ plugins/plugin-utils/project.json
@@ -1,0 +1,20 @@
+{
+  "name": "plugin-utils",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "plugins/plugin-utils/src",
+  "projectType": "library",
+  "targets": {
+    "build": {
+      "executor": "@nx/js:swc",
+      "inputs": ["{projectRoot}/src/**"],
+      "outputs": ["{options.outputPath}"],
+      "cache": true,
+      "options": {
+        "outputPath": "{projectRoot}/dist",
+        "main": "{projectRoot}/src/index.ts",
+        "tsConfig": "{projectRoot}/tsconfig.lib.json"
+      }
+    }
+  },
+  "tags": ["no-batching"]
+}
